packages/hate_speech_filter/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
grpcio==1.66.2
grpcio-tools==1.66.2
huggingface-hub==0.25.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy<2
packaging==24.1
protobuf==5.28.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
tokenizers==0.20.0
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.4
torch==2.3.1